+++ UVOCBot.Api/Migrations/20210312223201_UpdateMemberGroup.cs
@@ -1,0 +1,24 @@
+﻿using Microsoft.EntityFrameworkCore.Migrations;
+
+namespace UVOCBot.Api.Migrations
+{
+    public partial class UpdateMemberGroup : Migration
+    {
+        protected override void Up(MigrationBuilder migrationBuilder)
+        {
+            migrationBuilder.AddColumn<ulong>(
+                name: "CreatorId",
+                table: "MemberGroups",
+                type: "bigint unsigned",
+                nullable: false,
+                defaultValue: 0ul);
+        }
+
+        protected override void Down(MigrationBuilder migrationBuilder)
+        {
+            migrationBuilder.DropColumn(
+                name: "CreatorId",
+                table: "MemberGroups");
+        }
+    }
+}